+++ components/Match/Match.jsx
@@ -1,6 +1,11 @@
 import React from 'react';
-import { MatchTable } from '../Table';
+import { Table } from '../Table';
 import Error from '../Error';
+import { REDUCER_KEY } from '../../reducers';
+import { getMatch } from '../../actions';
+import { connect } from 'react-redux';
+import createConstantsWrapper from '../Constants';
+
 const columns = [{
   displayName: 'Hero',
   field: 'hero_id',
@@ -22,9 +27,33 @@ const columns = [{
   width: 2,
   sorted: false,
 }];
-export default ({ params }) => (
-  <div>
-    {params && params.match_id && <MatchTable matchId={params.match_id} columns={columns} />}
-    {(!params || !params.match_id) && <Error />}
-  </div>
-);
+
+const mapStateToProps = (state) => {
+  const { error, loading, match } = state[REDUCER_KEY].gotMatch;
+  return {
+    loading,
+    error,
+    data: match,
+  };
+};
+
+const mapDispatchToProps = (dispatch) => ({
+  sort: (column) => dispatch(getMatch(column)),
+  getMatch: (matchId) => dispatch(getMatch(matchId)),
+});
+
+class RequestLayer extends React.Component {
+  componentDidMount() {
+    this.props.getMatch(this.props.matchId);
+  }
+
+  render() {
+    return <div>
+      { <Table {...this.props} data={this.props.data && this.props.data.players ? this.props.data.players : this.props.data} columns={columns} /> }
+    </div>;
+  }
+}
+
+const TableWrapper = connect(mapStateToProps, mapDispatchToProps)(RequestLayer);
+
+export default createConstantsWrapper(TableWrapper);
