+++ components/Table/index.js
@@ -1,6 +1,5 @@
 import Table from './Table';
 import PlayerMatchesTable from './PlayerMatchesTable.container';
-import MatchTable from './MatchTable.container';
 
 export default Table;
-export { PlayerMatchesTable, MatchTable };
+export { PlayerMatchesTable };
